sql: split the SRF rewrite code to its own file
Release note: None
